moved hygro to submodule of core and adapted docs
