--- docs/source/api/hygro_api.md
@@ -1,22 +1,0 @@
----
-jupytext:
-  formats: md:myst
-  text_representation:
-    extension: .md
-    format_name: myst
-    format_version: 0.13
-    jupytext_version: 1.13.8
-kernelspec:
-  display_name: Python 3
-  language: python
-  name: pyrealm_python3
----
-
-
-# The {mod}`~pyrealm.hygro` module
-
-```{eval-rst}
-.. automodule:: pyrealm.hygro
-    :autosummary:
-    :members:
-```
