+++ docs/source/index.md
@@ -27,13 +27,9 @@ The `tmodel` module
 : Estimating plant allocation of gross primary productivity to growth and respiration,
   using the [T Model](users/tmodel/tmodel) {cite:p}`Li:2014bc`.
 
-The `hygro` module
-: Provides functions for [converting common hygrometric variables](users/hygro) to
-  vapour pressure deficit for use in the P Model.
-
 The `core` module
 : Contains fundamental utilities and physics functionality which is shared across the
-code base.
+code base. Contains the [hygro](users/hygro) and the utilities submodules.
 
 ## Indices and tables
