+++ docs/source/users/hygro.md
@@ -24,7 +24,7 @@ kernelspec:
 
 from matplotlib import pyplot
 import numpy as np
-from pyrealm import hygro
+from pyrealm.core import hygro
 %matplotlib inline
 
 # Set the resolution of examples
@@ -45,7 +45,8 @@ rh_2d = np.broadcast_to(rh_1d, (n_pts, n_pts))
 The {class}`~pyrealm.pmodel.pmodel_environment.PModelEnvironment` class requires
 vapour pressure
 deficit (VPD, Pa) as an input, but forcing datasets often provide alternative
-representations. The {mod}`~pyrealm.hygro`  provide functions to calculate saturated
+representations. The {mod}`~pyrealm.core.hygro`  provide functions to calculate
+saturated
 vapour pressure for a given temperature and the conversions from vapour pressure,
 relative humidity and specific humidity to vapour pressure deficit.
 
@@ -57,13 +58,13 @@ Deficit (VPD).  It is now usual for VP to be provided in kilopascals (kPa) but
 some older data sources use hectopascals (hPa), which are equivalent to millibars
 (mb or mbar).
 
-The function {func}`~pyrealm.hygro.convert_vp_to_vpd` takes values in kPa
+The function {func}`~pyrealm.core.hygro.convert_vp_to_vpd` takes values in kPa
 and returns kPa, so if you are using VP to prepare input data for
 {class}`~pyrealm.pmodel.pmodel_environment.PModelEnvironment`:
 
 * Make sure you are passing VP values to
-  {func}`~pyrealm.hygro.convert_vp_to_vpd` in kPa and not hPa or mbar.
-* Rescale the output of {func}`~pyrealm.hygro.convert_vp_to_vpd` from
+  {func}`~pyrealm.core.hygro.convert_vp_to_vpd` in kPa and not hPa or mbar.
+* Rescale the output of {func}`~pyrealm.core.hygro.convert_vp_to_vpd` from
   kPa to Pa, before using it in
   {class}`~pyrealm.pmodel.pmodel_environment.PModelEnvironment`.
