+++ pyrealm/core/__init__.py
@@ -1,5 +1,5 @@
 """The core module.
 
 Contains fundamental utilities and physics functionality which is shared across the
-code base.
+code base. Submodules are the utilities and the hygro submodule.
 """
