+++ pyrealm/core/hygro.py
@@ -1,8 +1,8 @@
-"""The :mod:`~pyrealm.hygro` module provides provides conversion functions for common
-hygrometric variables. The module provides conversions to vapour pressure deficit, which
-is the required input for the :class:`~pyrealm.pmodel.pmodel.PModel` from vapour
-pressure, specific humidity and relative humidity. The implementation is drawn from and
-validated against the ``bigleaf`` R package.
+"""The :mod:`~pyrealm.core.hygro` submodule provides provides conversion functions for
+common hygrometric variables. The module provides conversions to vapour pressure
+deficit, which is the required input for the :class:`~pyrealm.pmodel.pmodel.PModel`
+from vapour pressure, specific humidity and relative humidity. The implementation is
+drawn from and validated against the ``bigleaf`` R package.
 """  # noqa: D205, D415
 
 import numpy as np
